--- src/main/java/com/hello/boilerplate/domain/user/dto/UserRequestDto.java
@@ -1,17 +1,0 @@
-package com.hello.boilerplate.domain.user.dto;
-
-import com.hello.boilerplate.domain.user.domain.Role;
-import com.hello.boilerplate.domain.user.domain.User;
-
-public class UserRequestDto {
-    public record Register(
-            String loginId,
-            String password,
-            String email,
-            String name
-    ) {
-        public User toEntity(final String encodedPassword) {
-            return new User(loginId, encodedPassword, email, name, Role.USER);
-        }
-    }
-}
